server/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
beautifulsoup4
aiohttp
openai
numpy
scikit-learn
transformers
torch